reply_challenge/pyproject.toml
[project]
name = "reply_challenge"
version = "0.1.0"
description = "reply-challenge using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]==1.14.1",
    "langfuse>=3.0.0,<4.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "openai-whisper>=20250625",
    "ulid-py>=1.1.0,<2.0.0"
]

[project.scripts]
reply_challenge = "reply_challenge.main:run"
run_crew = "reply_challenge.main:run"
train = "reply_challenge.main:train"
replay = "reply_challenge.main:replay"
test = "reply_challenge.main:test"
run_with_trigger = "reply_challenge.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
